+++ packages/design-system/src/components/Dropdown/Dropdown.tsx
@@ -30,6 +30,7 @@ type DropdownButtonType = Omit<ClickableProps, 'children'> & {
 type DropdownLinkType = Omit<LinkableType, 'children'> & {
 	label: string;
 	type: 'link';
+	as: ReactElement;
 } & DataAttributes;
 
 type DropdownLabelType = {