+++ packages/design-system/src/components/Tabs/Primitive/TabsProvider.tsx
@@ -8,6 +8,7 @@ export type TabsProviderPropTypes = {
 	activeKey?: string;
 	onSelect?: (event: any, key: string) => void;
 	size?: string;
+	id?: string;
 };
 
 type WithChildren = {
@@ -29,7 +30,7 @@ export function TabsProvider(props: TabsProviderPropTypes & WithChildren) {
 		},
 	});
 	return (
-		<nav>
+		<nav id={props.id}>
 			<StackVertical gap="M">
 				<TabsInternalContext.Provider value={{ size: props.size, ...controlled }}>
 					{props.children}