+++ packages/storybook-docs/webpack.custom.js
@@ -20,9 +20,12 @@ module.exports = {
 				exclude: /node_modules/,
 				use: [
 					{ loader: MiniCssExtractPlugin.loader, options: { esModule: false } },
-					{ loader: 'css-loader', options: { sourceMap: true } },
 					{
-						loader: 'postcss-loader',
+						loader: require.resolve('css-loader'),
+						options: { sourceMap: true },
+					},
+					{
+						loader: require.resolve('postcss-loader'),
 						options: {
 							postcssOptions: {
 								plugins: ['autoprefixer'],
@@ -30,9 +33,12 @@ module.exports = {
 							},
 						},
 					},
-					{ loader: 'resolve-url-loader', options: { sourceMap: true } },
 					{
-						loader: 'sass-loader',
+						loader: require.resolve('resolve-url-loader'),
+						options: { sourceMap: true },
+					},
+					{
+						loader: require.resolve('sass-loader'),
 						options: { sourceMap: true },
 					},
 				],