+++ tools/scripts-config-react-webpack/config/icons.js
@@ -14,6 +14,14 @@ function getThemeIcon(theme) {
  */
 function getTalendIconsPath() {
 	const main = require.resolve('@talend/icons');
+
+	if (main.indexOf('.pnpm') > -1) {
+		const startPath = main.substring(0, main.indexOf('icons'));
+		const regex = /@talend\+icons@([^\s/]+)/;
+		const match = main.match(regex);
+		const version = match[1];
+		return `${startPath}icons@${version}/node_modules/@talend/icons`;
+	}
 	const root = main.split('icons')[0];
 	return `${root}icons`;
 }