+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @talend/ui-playground
 
+## 2.0.0
+
+### Major Changes
+
+- 4044f6198: ARCH-662 - Bump i18next from v20 to v23 and react-i18next from v11 to v13
+
+### Patch Changes
+
+- Updated dependencies [96d688489]
+- Updated dependencies [9a0732ec5]
+- Updated dependencies [da4cf2971]
+- Updated dependencies [4044f6198]
+  - @talend/react-faceted-search@12.0.0
+  - @talend/react-bootstrap@2.0.0
+  - @talend/react-cmf-router@6.0.0
+  - @talend/react-components@12.0.0
+  - @talend/react-containers@10.0.0
+  - @talend/react-dataviz@4.0.0
+  - @talend/react-forms@11.0.0
+  - @talend/icons@7.0.0
+  - @talend/react-cmf@8.0.0
+  - @talend/bootstrap-theme@8.2.1
+
 ## 1.1.15
 
 ### Patch Changes